requirements.txt
certifi==2020.6.20
charset-normalizer==2.0.12
cycler==0.11.0
idna==3.3
kiwisolver==1.3.1
lxml==4.8.0
matplotlib==3.3.4
multitasking==0.0.10
numpy==1.19.5
pandas==1.1.5
Pillow==8.4.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
six==1.16.0
urllib3==1.26.8
yfinance==0.1.70
